\name{daterangePicker}
\alias{daterangePicker}
\title{Date Range Piker}
\usage{
  daterangePicker(inputId, label)
}
\arguments{
  \item{inputId}{Specifies the input slot that will be used
  to access the value.}

  \item{label}{The contents of the button--usually a text
  label, but you could also use any other HTML, like an
  image.}
}
\description{
  Creates a dropdown menu from which a user can select a
  range of dates. This is based on
  https://github.com/dangrossman/bootstrap-daterangepicker.
}
\seealso{
  [bootstrap-daterangepicker](https://github.com/dangrossman/bootstrap-daterangepicker)
}

