\name{generateHTMLTableHeader}
\alias{generateHTMLTableHeader}
\title{Generate HTML Table <thead/>}
\usage{
  generateHTMLTableHeader(df)
}
\arguments{
  \item{df}{data.frame() to convert to a html table}
}
\value{
  HTML <thead/>
}
\description{
  Helper function which generates the html <thead/> to wrap
  the given data.frame()'s column names
}
\examples{
generateHTMLTableHeader(df)
}

