\name{htmlTableHead}
\alias{htmlTableHead}
\title{htmlTableHead}
\usage{
  htmlTableHead(dataframe, include.rownames = TRUE)
}
\arguments{
  \item{dataframe}{A data frame to generate HTML table}

  \item{include.rownames}{Include rownames in generated
  table?. Default: TRUE

  Generate HTML Table Head component}
}
\description{
  htmlTableHead
}

