\name{navbarMenu}
\alias{navbarMenu}
\title{Create a navbar input control}
\usage{
  navbarMenu(inputId)
}
\arguments{
  \item{inputId}{Input variable to assign the control's
  value to}
}
\value{
  A navbar menu that can be added to a UI definition.
}
\description{
  Create an input control for navigation bar menu
}

