\name{navbarPanel}
\alias{navbarPanel}
\title{Create a top navbar input panel}
\usage{
  navbarPanel(inputId)
}
\arguments{
  \item{inputId}{Input variable to assign the control's
  value to}
}
\value{
  A navbar menu that can be added to a UI definition.
}
\description{
  Create an input control for a top navigation bar menu
  panel
}

