\name{shinyExt-package}
\alias{shinyExt-package}
\alias{shinyExt}
\docType{package}
\title{
shinyExp
}
\description{
Add Extra functionalities from jQuery and Twitter Bootstrap for shiny.
}
\details{
\tabular{ll}{
Package: \tab shinyExt\cr
Type: \tab Package\cr
Version: \tab 0.0.1\cr
Date: \tab 2012-11-21\cr
License: \tab GPL-3\cr
}
Use in the same way like shiny package
}
\author{
Marcio Nicolau

Maintainer: Marcio Nicolau <marcionicolau@me.com>
}
\references{
jQuery
}
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[shiny:shiny-package]{shiny}} ~~
}
\examples{

}
